+++ chrome/browser/ui/views/task_manager_view_browsertest.cc
@@ -115,7 +115,7 @@ class TaskManagerViewTest : public InProcessBrowserTest {
     SessionID tab_id = sessions::SessionTabHelper::IdForTab(tab);
     std::unique_ptr<TaskManagerTester> tester =
         TaskManagerTester::Create(base::RepeatingClosure());
-    for (size_t i = 0; i < static_cast<size_t>(tester->GetRowCount()); ++i) {
+    for (size_t i = 0; i < tester->GetRowCount(); ++i) {
       if (tester->GetTabId(i) == tab_id)
         return i;
     }
@@ -282,7 +282,7 @@ IN_PROC_BROWSER_TEST_F(TaskManagerViewTest, DISABLED_SelectionConsistency) {
   std::unique_ptr<TaskManagerTester> tester =
       TaskManagerTester::Create(base::RepeatingClosure());
   std::vector<content::WebContents*> tabs;
-  for (int i = 0; i < tester->GetRowCount(); ++i) {
+  for (size_t i = 0; i < tester->GetRowCount(); ++i) {
     // Filter based on our title.
     if (!base::MatchPattern(tester->GetRowTitle(i), pattern))
       continue;
