+++ ui/views/controls/table/test_table_model.h
@@ -11,7 +11,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 class TestTableModel : public ui::TableModel {
  public:
-  explicit TestTableModel(int row_count);
+  explicit TestTableModel(size_t row_count);
 
   TestTableModel(const TestTableModel&) = delete;
   TestTableModel& operator=(const TestTableModel&) = delete;
@@ -19,13 +19,13 @@ class TestTableModel : public ui::TableModel {
   ~TestTableModel() override;
 
   // ui::TableModel overrides:
-  int RowCount() override;
-  std::u16string GetText(int row, int column_id) override;
-  ui::ImageModel GetIcon(int row) override;
+  size_t RowCount() override;
+  std::u16string GetText(size_t row, int column_id) override;
+  ui::ImageModel GetIcon(size_t row) override;
   void SetObserver(ui::TableModelObserver* observer) override;
 
  private:
-  int row_count_;
+  size_t row_count_;
   raw_ptr<ui::TableModelObserver> observer_;
 };
